fix: Richtext components with Markdown support and improved styling
- Migrate RichText components to a common UI location
- Add Markdown support using tiptap-markdown extension
- Improve editor and content rendering with responsive styles
- Update package.json with new dependencies (lodash, tiptap-markdown)
- Reorganize and optimize RichText component structure
