+++ frontend/package.json
@@ -1,46 +1,49 @@
 {
-  "name": "frontend_type",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "phone": "vite --host 0.0.0.0 --mode phone",
-    "build": "tsc -p tsconfig.build.json && vite build",
-    "preview": "vite preview",
-    "generate-client": "openapi-ts"
-  },
-  "dependencies": {
-    "@chakra-ui/react": "^3.8.0",
-    "@emotion/react": "^11.14.0",
-    "@lottiefiles/dotlottie-react": "^0.13.0",
-    "@tanstack/react-query": "^5.66.0",
-    "@tanstack/react-router": "^1.102.1",
-    "@tanstack/router-vite-plugin": "^1.102.1",
-    "@tiptap/extension-character-count": "^2.11.5",
-    "@tiptap/pm": "^2.11.5",
-    "@tiptap/react": "^2.11.5",
-    "@tiptap/starter-kit": "^2.11.5",
-    "@vitejs/plugin-react-swc": "^3.8.0",
-    "axios": "^1.7.9",
-    "next-themes": "^0.4.4",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "react-hook-form": "^7.54.2",
-    "react-icons": "^5.4.0"
-  },
-  "devDependencies": {
-    "@biomejs/biome": "^1.9.4",
-    "@hey-api/openapi-ts": "^0.57.0",
-    "@tanstack/router-devtools": "^1.102.1",
-    "@tanstack/router-plugin": "^1.102.1",
-    "@types/node": "^22.13.1",
-    "@types/react": "^19.0.8",
-    "@types/react-dom": "^19.0.3",
-    "@vitejs/plugin-react": "^4.3.4",
-    "globals": "^15.14.0",
-    "typescript": "~5.7.2",
-    "vite": "^6.1.0",
-    "vite-tsconfig-paths": "^5.1.4"
-  }
+	"name": "frontend_type",
+	"private": true,
+	"version": "0.0.0",
+	"type": "module",
+	"scripts": {
+		"dev": "vite",
+		"phone": "vite --host 0.0.0.0",
+		"build": "tsc -p tsconfig.build.json && vite build",
+		"preview": "vite preview",
+		"generate-client": "openapi-ts"
+	},
+	"dependencies": {
+		"@chakra-ui/react": "^3.8.0",
+		"@emotion/react": "^11.14.0",
+		"@lottiefiles/dotlottie-react": "^0.13.0",
+		"@tanstack/react-query": "^5.66.0",
+		"@tanstack/react-router": "^1.102.1",
+		"@tanstack/router-vite-plugin": "^1.102.1",
+		"@tiptap/extension-character-count": "^2.11.5",
+		"@tiptap/pm": "^2.11.5",
+		"@tiptap/react": "^2.11.5",
+		"@tiptap/starter-kit": "^2.11.5",
+		"@vitejs/plugin-react-swc": "^3.8.0",
+		"axios": "^1.7.9",
+		"lodash": "^4.17.21",
+		"next-themes": "^0.4.4",
+		"react": "^19.0.0",
+		"react-dom": "^19.0.0",
+		"react-hook-form": "^7.54.2",
+		"react-icons": "^5.4.0",
+		"tiptap-markdown": "^0.8.10"
+	},
+	"devDependencies": {
+		"@biomejs/biome": "^1.9.4",
+		"@hey-api/openapi-ts": "^0.57.0",
+		"@tanstack/router-devtools": "^1.102.1",
+		"@tanstack/router-plugin": "^1.102.1",
+		"@types/lodash": "^4.17.16",
+		"@types/node": "^22.13.1",
+		"@types/react": "^19.0.8",
+		"@types/react-dom": "^19.0.3",
+		"@vitejs/plugin-react": "^4.3.4",
+		"globals": "^15.14.0",
+		"typescript": "~5.7.2",
+		"vite": "^6.1.0",
+		"vite-tsconfig-paths": "^5.1.4"
+	}
 }
