+++ frontend/src/components/commonUI/RichText/RichTextContent.tsx
@@ -2,7 +2,8 @@ import { Box } from "@chakra-ui/react";
 import { EditorContent, useEditor } from "@tiptap/react";
 import StarterKit from "@tiptap/starter-kit";
 import { useEffect, useRef, useState } from "react";
-import "@/components/shared/RichText.styles.css";
+import { Markdown } from "tiptap-markdown";
+import "@/components/commonUI/RichText/RichText.styles.css";
 
 interface RichTextContentProps {
 	content: string;
@@ -18,10 +19,24 @@ export default function RichTextContent({
 	const [shouldCenter, setShouldCenter] = useState(true);
 
 	const editor = useEditor({
-		extensions: [StarterKit],
+		extensions: [
+			StarterKit,
+			Markdown.configure({
+				html: false,
+				transformPastedText: true,
+				transformCopiedText: true,
+				breaks: true,
+			}),
+		],
 		content,
 		editable: false,
 	});
+
+	useEffect(() => {
+		if (editor && content) {
+			editor.commands.setContent(content);
+		}
+	}, [content, editor]);
 
 	useEffect(() => {
 		if (!isVisible) return;
@@ -36,7 +51,7 @@ export default function RichTextContent({
 		const containerHeight = container.clientHeight;
 		const contentHeight = content.scrollHeight;
 
-		if (contentHeight <= containerHeight * 0.25) {
+		if (contentHeight <= containerHeight * 0.15) {
 			const multiplier = containerHeight / contentHeight;
 			if (multiplier > 1) {
 				const newMultiplier = Math.min(multiplier, 1.5);
